configured browser stack changes

+++ src/test/java/com/om/framework/basetest/BaseTest.java
@@ -46,7 +46,7 @@ public class BaseTest extends TestListenerAdapter
 	public static String sChromeExtension="";
 	public static String sFfDriverPath = "";
 	public static String sNewParameter = "";
-		 	
+
 	//Test Case members
 	public static String Subject = "";
 	public static String TestName = "";
@@ -55,27 +55,36 @@ public class BaseTest extends TestListenerAdapter
 	public static String ExpectedResult = "";	
 	public static String Type = "";	
 	public static String TestDate = "";
-	
+
 	//public static String testCaseFilename = "";
 	private static Calendar cal = Calendar.getInstance();
 	private static SimpleDateFormat sdf = new SimpleDateFormat("MMddyyHHmmss");
 	private static String timeSubStr = sdf.format(cal.getTime());
 	public static String testCaseFilename = "\\TC_WebClient_TestCase-"+timeSubStr+".csv";
-		
+
 	//Test Logging members
 	public static String sTestPlanLogging = "";
 	public static String dbplatform = "";
 	public static boolean imgCaptured = false;
-	
+
 	//Language setting member
 	protected static String slxLanguage = "";
-	
+
 	public static Map<String, String> objProperties = null;
-	
+
 	public static FirefoxProfile ffProfile = new FirefoxProfile();
 	public static WebDriver driver;
-	
-	
+
+
+	//Browserstack members
+	public static String sbStackUserName="";
+	public static String sbStackpasskey="";
+	public static String sbStackOS="";
+	public static String sbStackOs_version="";
+	public static String sbStackBrowser="";
+	public static String sbStackBrowser_version="";
+	public static String sbStackSelenium_version="";
+
 	/**
 	 * Launches the test webdriver browser based on the given values in the app.properties or incoming input values
 	 * from the command-line or TestNG .xml.  If the app.properties does not override the input values or if the input 
@@ -104,7 +113,7 @@ public class BaseTest extends TestListenerAdapter
 		try {
 			//read properties file and set local variable values (corresponding non-null values will override the input values)
 			readPropertiesFileNSetLocalVars();
-			
+
 			//process input parameter values (if not overridden in app.properties)
 			if (sBrowser.equals("") || sBrowser.equals(null)) {
 				sBrowser = inputBrowser;
@@ -122,7 +131,7 @@ public class BaseTest extends TestListenerAdapter
 				sPassword = inputPassword;
 				System.out.println("> sPassword parameter set to: " + sPassword);
 			}
-						
+
 
 			//open the appropriate browser session
 			if(sBrowser.equalsIgnoreCase("IE"))
@@ -131,6 +140,8 @@ public class BaseTest extends TestListenerAdapter
 				driver = Browser.openBrowser(sBrowser, sTestSiteURL, sEdgeDriverPath);
 			else if(sBrowser.equalsIgnoreCase("chrome") || sBrowser.equalsIgnoreCase("cr") || sBrowser.equalsIgnoreCase("chromeHeadless"))
 				driver = Browser.openBrowser(sBrowser, sTestSiteURL, sChromeDriverPath);
+			else if(sBrowser.equalsIgnoreCase("Browserstack"))
+				driver = Browser.openBrowserStack(sTestSiteURL,sbStackUserName,sbStackpasskey,sbStackOS,sbStackOs_version,sbStackBrowser,sbStackBrowser_version,sbStackSelenium_version);
 			else if(sBrowser.equalsIgnoreCase("chrome-x"))
 				driver = Browser.openChromeBrowserWithXtension(sTestSiteURL, sChromeDriverPath, sChromeExtension);
 			else if((sBrowser.equalsIgnoreCase("ff") || sBrowser.equalsIgnoreCase("firefox")) && sFfDriverPath!=null)
@@ -142,19 +153,19 @@ public class BaseTest extends TestListenerAdapter
 				Messages.errorMsg="No browser drivers found";
 				return;
 			}
-			
+
 			//set the default wait and timeout values 
-			driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
-			driver.manage().timeouts().pageLoadTimeout(200, TimeUnit.SECONDS);						
+			//driver.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
+			//driver.manage().timeouts().pageLoadTimeout(200, TimeUnit.SECONDS);						
 		} 
 		catch (Exception e) 
 		{
 			e.printStackTrace();
 		}
-		
+
 	}
-	
-	
+
+
 	/**
 	 * Reads the app.properties file from the project's working directory and assigns the values to local variables.
 	 * 
@@ -172,14 +183,14 @@ public class BaseTest extends TestListenerAdapter
 		try
 		{
 			objProperties = new HashMap<String, String>();
-			
+
 			reader = new FileReader("app.properties");
 			p.load(reader);
 			reader.close();
-			
+
 			//get all the keys defined in properties file
 			Set<Object> keys = p.keySet();
-			
+
 			//get each key and value. Put them in a hash map.
 			for(Object k:keys)
 			{
@@ -187,7 +198,7 @@ public class BaseTest extends TestListenerAdapter
 				sValue =  p.getProperty(sKey);
 				objProperties.put(sKey, sValue);
 			}
-						
+
 			// these variables are not required but still keeping them as they are very commonly used one.
 			//If you define any new property in app.properties, that can be directly taken from the BaseTest.objProperties.get("property");
 			sTestSiteURL = p.getProperty("test_url");
@@ -195,30 +206,39 @@ public class BaseTest extends TestListenerAdapter
 			sPassword = p.getProperty("password");
 			sBrowser = p.getProperty("browser");
 			sLanguage = p.getProperty("language");
-			
+
 			sProjectName = p.getProperty("project_name");
 			sBuild = p.getProperty("build");
 			sDBPlatForm = p.getProperty("db_platform");
 			sTestCasePath = p.getProperty("testcase_path");
-			
+
 			sIEDriverPath = p.getProperty("iedriverpath");
 			sEdgeDriverPath = p.getProperty("edgedriverpath");
 			sChromeDriverPath = p.getProperty("chromedriverpath");
 			sFfDriverPath = p.getProperty("ffdriverpath");
 			sChromeExtension = p.getProperty("chromeextension");	
-			
+
 			sAdminUser = p.getProperty("adminuser");
 			sAdminPW = p.getProperty("adminpw");
-			
+
 			sTestPlanLogging = p.getProperty("logging");
+
+			//browserStack members
+			sbStackUserName=p.getProperty("bStackUserName");
+			sbStackpasskey=p.getProperty("bStackpasskey");
+			sbStackOS=p.getProperty("bStackOS");
+			sbStackOs_version=p.getProperty("bStackOs_version");
+			sbStackBrowser=p.getProperty("bStackBrowser");
+			sbStackBrowser_version=p.getProperty("bStackBrowser_version");
+			sbStackSelenium_version=p.getProperty("bStackSelenium_version");
 		}
 		catch(Exception e)
 		{
 			e.printStackTrace();
 		}
 	}
-	
-	
+
+
 	/**
 	 * Closes the WebDriver browser after the test suite is complete.
 	 * 
@@ -228,16 +248,23 @@ public class BaseTest extends TestListenerAdapter
 	@AfterClass(alwaysRun = true)
 	public static void closeBrowser() throws IOException {
 		String methodID = "closeBrowser";
-			
-		driver.close();
-		if (!sBrowser.equalsIgnoreCase("ff") && !sBrowser.equalsIgnoreCase("firefox")) { 
-			Browser.closeCurrentBrowser(driver);
+		if(sBrowser.equalsIgnoreCase("browserstack")) {
+			driver.quit();
+			System.out.println("Browserstack closed successfully");
 		}
-		else {
-			Runtime.getRuntime().exec("taskkill /T /F /IM geckodriver.exe");
+		else
+		{
+			driver.close();
+			if (!sBrowser.equalsIgnoreCase("ff") && !sBrowser.equalsIgnoreCase("firefox")) { 
+				Browser.closeCurrentBrowser(driver);
+			}
+			else {
+				Runtime.getRuntime().exec("taskkill /T /F /IM geckodriver.exe");
+			}
 		}
+
 		System.out.println(methodID + ": the WebDriver '" + sBrowser + "' browser was closed.");
-		
+
 		System.out.println("Test End Date & Time: " + Utilities.generateRandString());
 		System.out.println("");
 	}
